improvement(BaseNode): add scylla_sysconfig_setup
Add the procedure for running scylla_sysconfig_setup
script on the db node by using remoter commands.
The procedure involves copying the previous config
files on the db node, running the script, comparing
the new config files and if there is a difference,
draining and restarting the node (which might involve
a resharding).
New ContinousEvent is added to keep track of these
procedures and their runtime.