refactor(fullscan): add random generator to partition op class
Introducing a random generator with seeding to the fullscans. It
replaces the previous naked random method calls, so that we may
have reproducibility of fullscans when using the same seed value.

The logic for getting the sequence of fullscan operations is
similar to that of the nemesis module: we prepopulate a queue with
instances of fullscan operation classes and then, when running the
fullscan thread, we pop from that queue.

Both the queue population and logic for creating random fullscan
query commands uses the same random generator object, so that the
commands are reproduced in line with the fullscan operations sequence.